measurem with error

+++ Khayati.Core/DTO/Measurement/MeasurementAddDto.cs
@@ -5,18 +5,18 @@ namespace Khayati.Core.DTO.Measurement
     public class MeasurementAddDto
     {
         //public int CustomerId { get; set; }
-        public double Height { get; set; }
-        public double Chest { get; set; }
-        public double Waist { get; set; }
-        public double Leg { get; set; }
-        public double trousers { get; set; }
-        public double Neck { get; set; }
-        public double Sleeve { get; set; }
-        public double DateCreated { get; set; }
-        public double DateTaken { get; set; }
+        public int Height { get; set; }
+        public int Chest { get; set; }
+        public int Waist { get; set; }
+        public int Leg { get; set; }
+        public int trousers { get; set; }
+        public int Neck { get; set; }
+        public int Sleeve { get; set; }
+        public int DateCreated { get; set; }
+        public int DateTaken { get; set; }
 
-        public double ShoulderWidth { get; set; }
-        public double ArmLength { get; set; }
+        public int ShoulderWidth { get; set; }
+        public int ArmLength { get; set; }
         public int CustomerId { get; set; }
